+++ CHANGELOG.rst
@@ -6,6 +6,8 @@ Since last release
 ====================
 
 **Added:**
+
+* Added a warning for when a facility trades with itself (#1895)
 * Added a new datatype to the backend for tariff region (#1922)
 * Added Functions to access region and institution from facilities and institutions (#1914)
 * Added a unit value to Resource, Material, and Product objects (#1894)